Fix quote handling in String and Character literals
Previously, because character and string literals shared the
same lexing escape code, you could put a \' inside a String
literal and a \" inside a character literal. Both of these
should have been disallowed as they aren't valid escapes in that
context.

This commit adds an additional flag to escape to indicate if we
are executing in the context of a "" literal or not. We then
allow/disallow \' and \" as appropriate.

Fixes #760